badge_server/requirements.txt
compatibility_lib==0.0.10
Flask==1.0.2
google-cloud-bigquery==1.3.0
google-cloud-datastore==1.7.0
grpcio==1.15.0
pybadges==1.0.1
redis==2.10.5
requests==2.19.1
retrying==1.3.3
wrapt==1.10.11
